+++ frontend/js/components/matchHistory/matchHistory.js
@@ -10,3 +10,43 @@ export async function matchHistoryComponent() {
         console.error('Error fetching matchHistory:', error);
     }
 }
+
+export async function displayMatchHistory(userStatJson) {
+    let matchHistoryContainer = document.getElementById('matchHistory');
+    let matchHistory = await matchHistoryComponent();
+    matchHistoryContainer.appendChild(matchHistory);
+
+    let matchHistoryWinner = document.getElementById('winnerUsername');
+    let matchHistoryLoser = document.getElementById('loserUsername');
+    let matchHistoryWinScore = document.getElementById('winnerScore');
+    let matchHistoryLoseScore = document.getElementById('loserScore');
+
+    userStatJson.played_matches.forEach((game) => {
+        const listElement = document.createElement('li');
+        listElement.classList.add('list-group-item');
+        listElement.classList.add('border-0');
+        listElement.innerHTML = game.winner_username;
+        matchHistoryWinner.appendChild(listElement);
+    });
+    userStatJson.played_matches.forEach((game) => {
+        const listElement = document.createElement('li');
+        listElement.classList.add('list-group-item');
+        listElement.classList.add('border-0');
+        listElement.innerHTML = game.winner_score;
+        matchHistoryWinScore.appendChild(listElement);
+    });
+    userStatJson.played_matches.forEach((game) => {
+        const listElement = document.createElement('li');
+        listElement.classList.add('list-group-item');
+        listElement.classList.add('border-0');
+        listElement.innerHTML = game.loser_username;
+        matchHistoryLoser.appendChild(listElement);
+    });
+    userStatJson.played_matches.forEach((game) => {
+        const listElement = document.createElement('li');
+        listElement.classList.add('list-group-item');
+        listElement.classList.add('border-0');
+        listElement.innerHTML = game.loser_score;
+        matchHistoryLoseScore.appendChild(listElement);
+    });
+}
